+++ labbook_BE/default_settings.py
@@ -1,6 +1,6 @@
 # -*- coding:utf-8 -*-
 
-APP_VERSION = '3.3.13'
+APP_VERSION = '3.3.14'
 
 # ######################################
 #  Database parameters